EditPieceForm integrated with AddPieceForm
