+++ src/components/art/Art.js
@@ -1,8 +1,8 @@
-import { Link } from "react-router-dom"
+import { Link, useNavigate } from "react-router-dom"
 import "./art.css"
 
 export const Art = ({ art_piece, token }) => {
-
+    const navigate = useNavigate()
     return (
         <>
             <div class="tile is-parent is-4">
@@ -18,6 +18,9 @@ export const Art = ({ art_piece, token }) => {
                     <div>{art_piece.arttype?.label} {art_piece.media?.label}</div>
                     <div>{art_piece.length}'x{art_piece.width}'x{art_piece.height}'</div>
                     <div>${art_piece.price}</div>
+                    <button className="button is-rounded" onClick={() => navigate(`${art_piece.id}/edit`)}>
+                    Edit
+                </button>
                 </article>
             </div>
