+++ src/views/ApplicationViews.js
@@ -5,7 +5,7 @@ import { Portfolio } from "../components/art/Portfolio"
 import { Login } from "../components/auth/Login"
 import { Register } from "../components/auth/Register"
 import { Authorized } from "./Authorized"
-import { AddPieceForm } from "../components/art/AddPieceForm"
+import { AddPieceForm } from "../components/art/PieceForm"
 
 export const ApplicationViews = ({ token, setToken }) => {
   return (
@@ -23,6 +23,7 @@ export const ApplicationViews = ({ token, setToken }) => {
         <Route path="/portfolio" >
             <Route index element={<Portfolio/>} />
             <Route path="add" element={<AddPieceForm />} />
+            <Route path=":pieceId/edit" element={<AddPieceForm />} />
         </Route>
         <Route element={<Authorized token={token} />}>
           {/* Add Routes here */}
